- move lora files from hwconf/ to lora/ - modify Makefile and add lora/lora.mk - hopefully have all idents fixed - restore commands.c and add new function again - fix memory leak bug in gesc.c - rfm95w.c add delay in while loop - add gesc to build_all / rebuild all works
