+++ conf_general.h
@@ -24,7 +24,7 @@
 #define FW_VERSION_MAJOR			5
 #define FW_VERSION_MINOR			03
 // Set to 0 for building a release and iterate during beta test builds
-#define FW_TEST_VERSION_NUMBER		14
+#define FW_TEST_VERSION_NUMBER		16
 
 #include "datatypes.h"
 
@@ -209,7 +209,6 @@
  * Set APP_CUSTOM_TO_USE to the name of the main C file of the custom application.
  */
 //#define APP_CUSTOM_TO_USE			"app_custom_template.c"
-#define APP_CUSTOM_TO_USE			"app_dpv.c"
 //#define APP_CUSTOM_TO_USE			"app_motor_heater.c"
 //#include "app_erockit_conf_v2.h"
 //#include "finn/app_finn_az_conf.h"
