+++ hwconf/hwconf.mk
@@ -3,9 +3,6 @@ HWSRC =	hwconf/hw.c \
 	hwconf/drv8305.c \
 	hwconf/drv8320s.c \
 	hwconf/drv8323s.c \
-    hwconf/SX1278.c \
-    hwconf/rfm95w.c \
-    hwconf/SX1278_hw.c \
 	hwconf/si8900.c
 
 HWINC = hwconf
