+++ lora/lora.mk
@@ -1,0 +1,5 @@
+LORASRC =   lora/SX1278.c \
+            lora/rfm95w.c \
+            lora/SX1278_hw.c 
+
+LORAINC = 	lora
